requirements.txt
# PyTorch with CUDA 12.4
torch==2.4.1 --extra-index-url https://download.pytorch.org/whl/cu124
torchvision==0.19.1 --extra-index-url https://download.pytorch.org/whl/cu124

# Data science and ML packages
pandas
tqdm
numpy
pillow
scikit-learn
opencv-python

# Hugging Face ecosystem
datasets
huggingface-hub
transformers
trl
bitsandbytes
peft
accelerate

# Web and API
gradio
litserve
fastapi
uvicorn

# Browser automation
selenium
webdriver-manager

# Monitoring and visualization
wandb
supervision
mss

# Utilities
num2words